requirements.txt
Django==2.2.28
redis==2.10.6
djangorestframework==3.11.2
djangorestframework-bulk==0.2.1
django-rest-swagger==2.2.0
django-cors-headers==3.2.1
mysqlclient==1.4.6
pyuwsgi==2.0.26
17MonIP==0.2.7
django-filter==2.3.0
gevent==24.2.1
python-consul2~=0.1.5
channels==2.4.0
daphne~=2.3
paramiko==3.3.2 # 2.7.2 2.8.1
channels-redis==3.0.1
XlsxWriter==1.3.6
xlrd==1.2.0
pyyaml==6.0.2
pyjwt==2.9.0
supervisor==4.2.5

# pip install gnureadline -i https://pypi.douban.com/simple/
